+++ tests/influence/torch/test_util.py
@@ -8,7 +8,6 @@ from pydvl.influence.torch.operator import MatrixOperator
 
 torch = pytest.importorskip("torch")
 import torch.nn  # noqa: F811
-from numpy.typing import NDArray
 from scipy.stats import pearsonr, spearmanr
 from torch.nn.functional import mse_loss
 from torch.utils.data import DataLoader, TensorDataset
@@ -24,12 +23,10 @@ from pydvl.influence.torch.util import (
     TorchLinalgEighException,
     TorchTensorContainerType,
     align_structure,
-    flatten_dimensions,
     safe_torch_linalg_eigh,
     torch_dataset_to_dask_array,
 )
 from tests.conftest import is_osx_arm64
-from tests.influence.conftest import linear_hessian_analytical, linear_model
 
 
 @dataclass
@@ -83,40 +80,6 @@ test_parameters = [
         regularization=0.00001,
     ),
 ]
-
-
-def linear_torch_model_from_numpy(A: NDArray, b: NDArray) -> torch.nn.Module:
-    """
-    Given numpy arrays representing the model $xA^t + b$, the function returns the corresponding torch model
-    :param A:
-    :param b:
-    :return:
-    """
-    output_dimension, input_dimension = tuple(A.shape)
-    model = torch.nn.Linear(input_dimension, output_dimension)
-    model.eval()
-    model.weight.data = torch.as_tensor(A, dtype=torch.get_default_dtype())
-    model.bias.data = torch.as_tensor(b, dtype=torch.get_default_dtype())
-    return model
-
-
-@pytest.fixture
-def model_data(request):
-    dimension, condition_number, train_size = request.param
-    A, b = linear_model(dimension, condition_number)
-    x = torch.rand(train_size, dimension[-1])
-    y = torch.rand(train_size, dimension[0])
-    torch_model = linear_torch_model_from_numpy(A, b)
-    vec = flatten_dimensions(
-        tuple(
-            torch.rand(*p.shape)
-            for name, p in torch_model.named_parameters()
-            if p.requires_grad
-        )
-    )
-    H_analytical = linear_hessian_analytical((A, b), x.numpy())
-    H_analytical = torch.as_tensor(H_analytical)
-    return torch_model, x, y, vec, H_analytical.to(torch.float32)
 
 
 @pytest.mark.torch
